moved CRUD operations for place images to a separate service and controller
